+++ src/app/(landing)/layout.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import '@/styles/landing.css'
+import './_styles/landing.css'
 
 interface LayoutProps {
   children: React.ReactNode
