+++ src/components/ui/form.tsx
@@ -11,7 +11,7 @@ import {
 } from 'react-hook-form'
 
 import { cn } from '@/utils'
-import { Label } from '@/components/ui/label'
+import Label from '@/components/ui/label'
 
 const Form = FormProvider
